* Change type of input/output * Add debug printf function

+++ bootloader/F1/Src/crypt.h
@@ -9,10 +9,10 @@ extern "C" {
 void extend_key(uint64_t kb[2 * ROUNDS], 
 	 	        const uint64_t K[2]);
 
-void encrypt(uint8_t* out, const uint8_t* in,
+void encrypt(uint64_t* out, const uint64_t* in,
              const uint32_t bytes, const uint64_t extended_key[2 * ROUNDS]);
 
-void decrypt(uint8_t* out, const uint8_t* in,
+void decrypt(uint64_t* out, const uint64_t* in,
              const uint32_t bytes, const uint64_t extended_key[2 * ROUNDS]);
 
 #ifdef __cplusplus
